Update Program.cs
Hi! I recommend to you to use StringBuilder instead of string, because string in C# are inmutable, and each time you add a letter to a string, you are making a new copy of it, and wasting a lot of memory. With StringBuilder it doesn´t happend, is a clean code, and don´t waste memory, and with Append method, you don´t need to convert to string each character.
Best of luck